split wallet test from mint test pipeline
