+++ tests/mint/test_mint_api_deprecated.py
