+++ tests/wallet/test_wallet_api.py
