+++ tests/wallet/test_wallet_auth.py
