+++ tests/wallet/test_wallet_cli.py
