+++ tests/wallet/test_wallet_restore.py
