+++ tests/wallet/test_wallet_subscription.py
